FTC-Core final push

+++ ForTheCrown-Core/src/main/java/net/forthecrown/core/events/SignShopDestroyEvent.java
@@ -1,6 +1,7 @@
 package net.forthecrown.core.events;
 
 import net.forthecrown.core.FtcCore;
+import net.forthecrown.core.exceptions.InvalidCommandExecution;
 import net.forthecrown.core.files.SignShop;
 import org.bukkit.ChatColor;
 import org.bukkit.block.Sign;
@@ -23,6 +24,11 @@ public class SignShopDestroyEvent implements Listener {
         } catch (Exception e){
             e.printStackTrace();
             return;
+        }
+
+        if(!shop.getOwner().equals(event.getPlayer().getUniqueId())){
+            event.setCancelled(true);
+            throw new InvalidCommandExecution(event.getPlayer(), "&cYou cannot destroy a sign you do not own!");
         }
 
         shop.destroyShop(); //destroys the shop
